Refactoring public header files into INTERFACE target
