+++ Gems/RecastNavigation/Code/Include/RecastNavigation/RecastHelpers.h
