+++ Gems/RecastNavigation/Code/Include/RecastNavigation/RecastSmartPointer.h
