+++ Gems/RecastNavigation/Code/recastnavigation_api_files.cmake
@@ -7,8 +7,11 @@
 #
 
 set(FILES
-    Include/RecastNavigation/RecastNavigationBus.h
+    Include/RecastNavigation/RecastHelpers.h
+    Include/RecastNavigation/RecastSmartPointer.h
+
     Include/RecastNavigation/DetourNavigationBus.h
+    Include/RecastNavigation/RecastNavigationBus.h
     Include/RecastNavigation/RecastNavigationMeshBus.h
     Include/RecastNavigation/RecastNavigationProviderBus.h
 )
